requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
assemblyai==0.17.0
anthropic==0.18.1
python-dotenv==1.0.1
pydantic==2.6.1